requirements.txt
pytrends
openai
requests
beautifulsoup4
python-dotenv
